Fix end-of-job slowdown for random IO with a random map

+++ filesetup.c
@@ -812,7 +812,7 @@ int init_random_map(struct thread_data *td)
 				(unsigned long long) td->o.rw_min_bs;
 		num_maps = (blocks + BLOCKS_PER_MAP - 1) /
 				(unsigned long long) BLOCKS_PER_MAP;
-		f->file_map = smalloc(num_maps * sizeof(int));
+		f->file_map = smalloc(num_maps * sizeof(unsigned long));
 		if (f->file_map) {
 			f->num_maps = num_maps;
 			continue;
